[IMP] Splitting model file into two one per model
