+++ ifrs_report/model/__init__.py
@@ -2,3 +2,4 @@
 
 from . import account
 from . import ifrs
+from . import ifrs_lines
